+++ src/models/buildinginfo.spec.ts
