+++ src/models/buildinginfo.ts
