+++ src/models/calendarEventClickedModel.ts
@@ -1,0 +1,6 @@
+import { CalendarEvent } from 'angular-calendar';
+
+export interface CalendarEventClicked {
+  event: CalendarEvent;
+  sourceEvent: MouseEvent;
+}
