+++ src/models/coordinates.spec.ts
