+++ src/models/directionModel.ts
@@ -1,0 +1,6 @@
+import { Coordinates } from './coordinates';
+
+export interface Direction {
+  origin: Coordinates;
+  destination: Coordinates;
+}
